--- Kanikama/Packages/net.shivaduke28.kanikama.bakery/Editor/Baking/Experimental.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 12a529d47d6746a08b2a470420474a8c
-timeCreated: 1687006712
