--- Kanikama/Packages/net.shivaduke28.kanikama.udon/Runtime/Experimental.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 3380cee698e74c38830cda5f11dd2fa2
-timeCreated: 1687033884
