+++ Kanikama/Packages/net.shivaduke28.kanikama.udon/Runtime/LTC/KanikamaUdonLTCUpdtater.asset
@@ -21,7 +21,7 @@ MonoBehaviour:
   compiledVersion: 2
   behaviourSyncMode: 2
   hasInteractEvent: 0
-  scriptID: 7530629844191762804
+  scriptID: -7430171427594678406
   serializationData:
     SerializedFormat: 2
     SerializedBytes:
