--- Kanikama/Packages/net.shivaduke28.kanikama/Editor/Baking/Experimental.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 67deb0645b1246f9b2d4c773c5158537
-timeCreated: 1686488512
