--- Kanikama/Packages/net.shivaduke28.kanikama/Runtime/Application/Experimental.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 3ea78c2877054058be5fc33e1ec754d6
-timeCreated: 1685797790
