+++ Kanikama/Packages/net.shivaduke28.kanikama/Runtime/Application/LTC/KanikamaLTCSurface.shader
@@ -64,7 +64,7 @@ Shader "Kanikama/KanikamaLTCSurface"
 
         #include <UnityPBSLighting.cginc>
         #include "Packages/net.shivaduke28.kanikama/Runtime/Application/Shaders/Kanikama.hlsl"
-        #include "Packages/net.shivaduke28.kanikama/Runtime/Application/Experimental/LTC/KanikamaLTC.hlsl"
+        #include "Packages/net.shivaduke28.kanikama/Runtime/Application/LTC/KanikamaLTC.hlsl"
 
         half4 _Color;
         half _Cutoff;
