--- Kanikama/Packages/net.shivaduke28.kanikama/Runtime/Baking/Experimental.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 7e340adfdae740118255f91ab5306934
-timeCreated: 1686488481
