--- Kanikama/Packages/net.shivaduke28.kanikama/Runtime/Baking/Experimental/LTC/Impl.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 59e5f55ff08c4aeba70f03c552e36960
-timeCreated: 1686486649
