+++ Kanikama/Packages/net.shivaduke28.kanikama/Runtime/Baking/Impl/LTC/KanikamaLTCMonitor.cs
@@ -1,6 +1,6 @@
 ﻿using UnityEngine;
 
-namespace Kanikama.Baking.Experimental.LTC.Impl
+namespace Kanikama.Baking.Impl.LTC
 {
     public sealed class KanikamaLTCMonitor : MonoBehaviour
     {
